Remove `not_past_last_scheduled_time? validation`
There is an issue with this gem where sometimes scheduled jobs
don't start, randomly.

The underlying problem here is that the gem polls jobs every minute
and checks if it's the correct time to run each job. This process sometimes
takes more than a minute to check all crons ending up not triggering them
since it's already past the cron time.

After investigating, the conclusion was that we can remove the
`not_past_scheduled_last_time?` validation and rely only on the
`not_enqueued_after?` because it already ensures that no job is
scheduled more than once.